Center feature section on large screens
